+++ packages/deno/packages/plugin-dataloader/schema-builder.ts
@@ -1,5 +1,6 @@
 // @ts-nocheck
-import SchemaBuilder, { InterfaceParam, ObjectParam, PothosSchemaError, SchemaTypes, ShapeFromTypeParam, } from '../core/index.ts';
+import type { GraphQLResolveInfo } from 'https://cdn.skypack.dev/graphql?dts';
+import SchemaBuilder, { InterfaceParam, ObjectParam, OutputRef, PothosSchemaError, SchemaTypes, ShapeFromTypeParam, } from '../core/index.ts';
 import { ImplementableLoadableNodeRef } from './refs/index.ts';
 import { ImplementableLoadableInterfaceRef } from './refs/interface.ts';
 import { ImplementableLoadableObjectRef } from './refs/object.ts';
@@ -61,7 +62,7 @@ schemaBuilderProto.loadableUnion = function loadableUnion<Key extends Dataloader
     return ref;
 };
 const TloadableNode = schemaBuilderProto.loadableNode;
-schemaBuilderProto.loadableNode = function loadableNode<Shape extends NameOrRef extends ObjectParam<SchemaTypes> ? ShapeFromTypeParam<SchemaTypes, NameOrRef, false> : object, Key extends DataloaderKey, Interfaces extends InterfaceParam<SchemaTypes>[], NameOrRef extends ObjectParam<SchemaTypes> | string, CacheKey = Key>(this: PothosSchemaTypes.SchemaBuilder<SchemaTypes>, nameOrRef: NameOrRef, options: LoadableNodeOptions<SchemaTypes, Shape, Key, Interfaces, NameOrRef, CacheKey>) {
+schemaBuilderProto.loadableNode = function loadableNode<Shape extends NameOrRef extends ObjectParam<SchemaTypes> ? ShapeFromTypeParam<SchemaTypes, NameOrRef, false> : object, Interfaces extends InterfaceParam<SchemaTypes>[], NameOrRef extends ObjectParam<SchemaTypes> | string, IDShape extends bigint | number | string = string, Key extends bigint | number | string = IDShape, CacheKey = Key>(this: PothosSchemaTypes.SchemaBuilder<SchemaTypes>, nameOrRef: NameOrRef, options: LoadableNodeOptions<SchemaTypes, Shape, Interfaces, NameOrRef, IDShape, Key, CacheKey>) {
     if (typeof (this as PothosSchemaTypes.SchemaBuilder<SchemaTypes> & Record<string, unknown>)
         .nodeInterfaceRef !== "function") {
         throw new PothosSchemaError("builder.loadableNode requires @pothos/plugin-relay to be installed");
@@ -73,13 +74,37 @@ schemaBuilderProto.loadableNode = function loadableNode<Shape extends NameOrRef 
         }).name ?? (nameOrRef as {
             name: string;
         }).name;
-    const ref = new ImplementableLoadableNodeRef<SchemaTypes, Shape, Shape, Key, CacheKey>(this, name, options);
+    const ref = new ImplementableLoadableNodeRef<SchemaTypes, Shape, Shape, IDShape, Key, CacheKey>(this, name, options);
     ref.implement({
         ...options,
         extensions: {
             ...options.extensions,
             pothosParseGlobalID: options.id.parse,
         },
+        isTypeOf: options.isTypeOf ??
+            (typeof nameOrRef === "function"
+                ? (maybeNode: unknown, context: object, info: GraphQLResolveInfo) => {
+                    if (!maybeNode) {
+                        return false;
+                    }
+                    if (maybeNode instanceof (nameOrRef as Function)) {
+                        return true;
+                    }
+                    const proto = Object.getPrototypeOf(maybeNode) as {
+                        constructor: unknown;
+                    };
+                    try {
+                        if (proto?.constructor) {
+                            const config = this.configStore.getTypeConfig(proto.constructor as OutputRef);
+                            return config.name === name;
+                        }
+                    }
+                    catch {
+                        // ignore
+                    }
+                    return false;
+                }
+                : undefined),
     });
     if (typeof nameOrRef !== "string") {
         this.configStore.associateRefWithName(nameOrRef, name);