+++ packages/deno/packages/plugin-dataloader/types.ts
@@ -55,4 +55,4 @@ export interface LoadableNodeId<Types extends SchemaTypes, Shape extends object,
         parse?: (id: string, ctx: Types["Context"]) => IDShape;
     };
 }
-export type LoadableNodeOptions<Types extends SchemaTypes, Shape extends object, Key extends bigint | number | string, Interfaces extends InterfaceParam<Types>[], NameOrRef extends ObjectParam<Types> | string, CacheKey> = DataloaderObjectTypeOptions<Types, Shape, Key, Interfaces, NameOrRef, CacheKey> & LoadableNodeId<Types, Shape, Key>;
+export type LoadableNodeOptions<Types extends SchemaTypes, Shape extends object, Interfaces extends InterfaceParam<Types>[], NameOrRef extends ObjectParam<Types> | string, IDShape extends bigint | number | string = string, Key extends bigint | number | string = IDShape, CacheKey = Key> = DataloaderObjectTypeOptions<Types, Shape, Key, Interfaces, NameOrRef, CacheKey> & LoadableNodeId<Types, Shape, IDShape>;