+++ packages/plugin-relay/tests/index.test.ts
@@ -513,6 +513,14 @@ describe('relay example schema', () => {
     it('parses ids', async () => {
       const query = gql`
         query {
+          idWithColon(id: "SURXaXRoQ29sb246MTp0ZXN0") {
+            id
+            idString
+          }
+          idsWithColon(ids: ["SURXaXRoQ29sb246MTp0ZXN0", "SURXaXRoQ29sb246Mjp0ZXN0OmV4YW1wbGU="]) {
+            id
+            idString
+          }
           numberThingByID(id: "TnVtYmVyOjE=") {
             id
             number
@@ -529,6 +537,16 @@ describe('relay example schema', () => {
             id
             number
           }
+          echoIDs(
+            globalID: "TnVtYmVyOjE="
+            numberThingID: "TnVtYmVyOjE="
+            genericNumberThingID: "TnVtYmVyOjE="
+          ) {
+            id
+            typename
+            arg
+            idType
+          }
         }
       `;
 
@@ -541,6 +559,40 @@ describe('relay example schema', () => {
       expect(result).toMatchInlineSnapshot(`
         {
           "data": {
+            "echoIDs": [
+              {
+                "arg": "globalID",
+                "id": "1",
+                "idType": "string",
+                "typename": "Number",
+              },
+              {
+                "arg": "numberThingID",
+                "id": "1",
+                "idType": "number",
+                "typename": "Number",
+              },
+              {
+                "arg": "genericNumberThingID",
+                "id": "1",
+                "idType": "string",
+                "typename": "Number",
+              },
+            ],
+            "idWithColon": {
+              "id": "SURXaXRoQ29sb246MTp0ZXN0",
+              "idString": "1:test",
+            },
+            "idsWithColon": [
+              {
+                "id": "SURXaXRoQ29sb246MTp0ZXN0",
+                "idString": "1:test",
+              },
+              {
+                "id": "SURXaXRoQ29sb246Mjp0ZXN0OmV4YW1wbGU=",
+                "idString": "2:test:example",
+              },
+            ],
             "invalid": null,
             "invalidList": null,
             "numberThingByID": {