+++ examples/open-telemetry/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pothos-examples/open-telemetry
 
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies [42bf6190]
+  - @pothos/core@3.27.0
+  - @pothos/plugin-tracing@0.5.7
+  - @pothos/tracing-opentelemetry@0.6.8
+
 ## 1.2.19
 
 ### Patch Changes